Moves next_awake tests into test_input
